% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcFile.R
\name{bcFile}
\alias{bcFile}
\title{Write a boundary condition file using a given dataframe.}
\usage{
bcFile(cFile, fileName, flowName, years, locations)
}
\arguments{
\item{cFile}{Send in a dataframe to be read into file.}

\item{fileName}{Name (and directory, if necessary) of output file}

\item{flowName}{Name (as a string) of the flow}

\item{years}{A string of the years your simulation will cover}

\item{locations}{AEM3D uses these numeric location tags to identify where the parameter affects the lake. Send in as a string double spaced.}
}
\description{
Write a boundary condition file using a given dataframe.
}
