% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{blend}
\alias{blend}
\title{Add flow as header "Q" to a data file.}
\usage{
blend(pFile, flows, col1, col2, col3)
}
\arguments{
\item{pFile}{dataframe containing the variables to be merged with the flow vector}

\item{flows}{vector containing flow values. Must be of equal length to the pFile, must match timestamps.}

\item{col1}{String name of the first column to be blended}

\item{col2}{String name of the second column to be blended}

\item{col3}{String name of the third column to be blended}
}
\value{
Returns a dataframe containing the three original vectors and the flow vector
}
\description{
Add flow as header "Q" to a data file.
}
