% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delQual.R
\name{delQual}
\alias{delQual}
\title{Delete unnecessary rows containing qualifiers.}
\usage{
delQual(pFile, qual, qualCol)
}
\arguments{
\item{pFile}{datafram containing the variables to be isolated}

\item{qual}{String of the qualifier used to denote removal in dataset}

\item{qualCol}{String name of the column containing qualifiers}
}
\value{
Returns the original dataframe with rows containing qualifiers removed.
}
\description{
Delete unnecessary rows containing qualifiers.
}
