% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate.R
\name{isolate}
\alias{isolate}
\title{Isolate columns from a .csv file}
\usage{
isolate(chemFile, col1, col2, col3, col4)
}
\arguments{
\item{chemFile}{Reads in a string denoting the filepath of your .csv}

\item{col1}{String name of an initial column}

\item{col2}{String name of a second column}

\item{col3}{String name of a third column}

\item{col4}{String name of a fourth column}
}
\value{
Returns a dataframe containing just the targetted columns
}
\description{
Function is intended to separate out useful data from
extraneous data that is unnecessary for AEM3D processing.
}
