% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder data with parameters as headers}
\usage{
reorder(pFile, col1, col2, col3)
}
\arguments{
\item{pFile}{dataframe containing variable information to be selected.}

\item{col1}{String name of the first column to be selected}

\item{col2}{String name of the second column to be selected}

\item{col3}{String name of the third column to be selected}
}
\value{
Returns a dataframe containing the three original vectors reordered around the first
}
\description{
Reorder data with parameters as headers
}
