% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLoadRun.R
\name{rLoadRun}
\alias{rLoadRun}
\title{Run rloadest with a given data frame.}
\usage{
rLoadRun(pFile, varName, flowName, dateName, unitsName, stationName)
}
\arguments{
\item{pFile}{The dataframe containing the variable, flow rate, and dates to be read}

\item{varName}{The string name of the variable to be run through rloadest}

\item{flowName}{The string name of the column containing flow data}

\item{dateName}{The string name of the column containing date data}

\item{unitsName}{A string of units denoting the concentration of the variable.}

\item{stationName}{The string name of the station.}
}
\value{
Returns a dataframe containing the concentration information
}
\description{
Run rloadest with a given data frame.
}
\seealso{
\code{\link[rloadest]{loadConvFactor}}
}
